% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_env.R
\name{setup_env}
\alias{setup_env}
\title{Setup environment variables in Verily workbench}
\usage{
setup_env(bucket = NULL, cdr = NULL, google_project = NULL)
}
\arguments{
\item{bucket}{optional character string to set WORKSPACE_BUCKET environment variable}

\item{cdr}{optional character string to set WORKSPACE_CDR environment variable}

\item{google_project}{optional character string to set GOOGLE_PROJECT environment variable}
}
\description{
Setup environment variables in Verily workbench
}
\details{
If any or all arguments are not provided, an attempt is made to discover the defaul environment variables.
}
\examples{
\dontrun{
# Try to automatically setup environment variables
setup_env()

# Set workspace bucket environment variable manually. Automatically set all others.
bucket <- "gs://..."
setup_env(bucket=bucket)
}
}
